endpoint por Id

+++ server/ProjectPokemon/ProjectPokemon/Models/Database/Repositories/NatureRepository.cs
@@ -1,7 +1,7 @@
 ﻿using ProjectPokemon.Models.Database.Entities;
 
 namespace ProjectPokemon.Models.Database.Repositories; 
-public class NatureRepository : BaseRepository<Nature, long> {
+public class NatureRepository : BaseRepository<Nature, int> {
     public NatureRepository(PokemonDbContext context) : base(context) {
     }
 }
